refactor: split detr.py → inference.py + variants.py; detr.py as re-export shim
Implements Item #8 from the design remediation plan (PR 6).

- Extract `ModelContext`, `_ModelContext`, `_build_model_context` into
  `src/rfdetr/inference.py` (new)
- Extract 14 concrete model variant classes into `src/rfdetr/variants.py`
  (new); documents the circular-import load-order invariant in the module
  docstring
- `src/rfdetr/detr.py` retains `RFDETR` base class and re-exports all
  extracted names at the bottom for full backward compatibility
- `rfdetr/__init__.py` is unchanged — all existing import paths continue
  to work through the `detr.py` shim
- Add `TestImportPaths` (22 tests) to `tests/utilities/test_package.py`
  verifying importability from all three paths and object identity across
  re-exports

detr.py: 865 → 622 lines
